third_party/WebKit/LayoutTests/fast/dom/shadow/multiple-content-combinator.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id="sandbox"></div>
<pre id="console"></pre>
<script>

function colorOf(node)
{
    return document.defaultView.getComputedStyle(node, null).getPropertyValue('color');
}

function cleanUp()
{
    document.getElementById('sandbox').innerHTML = '';
}

// This is the same test as content-pseudo-element-nested.html did.
// i.e. nested ::content => multiple /content/.
var sandbox = document.getElementById('sandbox');
sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('div', {'id': 'nested-host'},
                createShadowRoot(
                    createDOM('style', {},
                        document.createTextNode('div content /content/ div content /content/ div.green { color: green; }')),
                    document.createTextNode('text1'),
                    createDOM('div', {},
                        document.createTextNode('text2'),
                        createDOM('content', {'select': 'div'}))),
                createDOM('div', {'class': 'green'},
                    document.createTextNode('text3'),
                        createDOM('content', {})))),
        createDOM('div', {'class': 'green'},
            document.createTextNode('green')),
        createDOM('div', {},
            document.createTextNode('normal'))));

shouldBe('colorOf(document.querySelector("#host > div.green"))', '"rgb(0, 128, 0)"');

cleanUp();
</script>
</body>
</html>
