third_party/WebKit/LayoutTests/fast/dom/shadow/multiple-content-combinator-expected.txt

PASS colorOf(document.querySelector("#host > div.green")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
